requirements.txt
ultralytics>=8.3.20
torch>=2.1.0
PyYAML>=6.0
requests>=2.31.0
Pillow>=10.0.0
matplotlib>=3.8.0
super-gradients==3.6.0
pycocotools==2.0.6
onnxruntime==1.13.1
onnx==1.13.0
Flask>=3.0.0
